feat: add GOPRIVATE as a hardcoded environment variable
To use the go tool cm approach, we need to set GOPRIVATE so that
the go command is able to fetch (and cache the cm tool). This adds
the GOPRIVATE env variable to the generated .env file, which the
Justfile will read and thus will be able to run go tool cm without
much hassle.