refactor: runCommandWithOutput to allow more flexible JSON strings
This updates the runCommandWithOutput signature as follows:
- adds dir argument: the working directory from which to run the cmd
- switch to non-pointer return; it is awkward to work with pointers
  to slices, so it is better to return a value here. Especially,
  since both uses of this function passes a slice for T.

The main change, however, is that this now allows more flexibility
in how the output is processed if not immediately valid JSON.
Specifically, if the original output is not only JSON, then we
try to parse individual lines and check each line if it is valid
JSON. Each valid JSON line is then composed into a JSON array,
which is then returned.

This is in preparation for adding the gen-tests-json command.