feat: add new gen-tests-json command to generate tests.json file
This adds the new command gen-tests-json that will generate a
tests.json file in the labX folder based on the provided -labs
argument.

Note: If the tests fail, e.g., due to linter issues, the output
will not be returned. We should consider if this should be fixed,
or if we should fix require the user to fix the lint issues instead.